Extract CommandParameter to standard trait
