+++ renku-model/src/test/scala/io/renku/graph/model/entities/Generators.scala
@@ -22,9 +22,9 @@ import cats.syntax.all._
 import io.renku.generators.Generators.Implicits._
 import io.renku.generators.Generators.nonEmptyStrings
 import io.renku.graph.model.commandParameters._
-import io.renku.graph.model.testentities.CommandParameterBase.CommandInput._
-import io.renku.graph.model.testentities.CommandParameterBase.CommandOutput._
-import io.renku.graph.model.testentities.CommandParameterBase._
+import io.renku.graph.model.testentities.StepPlanCommandParameter.CommandInput._
+import io.renku.graph.model.testentities.StepPlanCommandParameter.CommandOutput._
+import io.renku.graph.model.testentities.StepPlanCommandParameter._
 import io.renku.graph.model.testentities.StepPlan.CommandParameters.CommandParameterFactory
 import io.renku.graph.model.testentities._
 import org.scalacheck.Gen
